-added x64 platform, no completely working right now...
git-svn-id: file:///root/svn_import/SolidSBC@2 bd7e3521-35e9-406e-9279-390287f868d3

+++ SolidSBCCliSvc/SolidSBCCliConfigSocket.cpp
@@ -74,7 +74,7 @@ bool CSolidSBCCliConfigSocket::WaitForConnect()
 	tv.tv_sec = 10; 
     tv.tv_usec = 0; 
 
-	int nErr = select(m_hCliConfSocket+1,NULL,&writefds,NULL,&tv);
+	int nErr = select((int)m_hCliConfSocket+1,NULL,&writefds,NULL,&tv);
 	if ( FD_ISSET(m_hCliConfSocket,&writefds) ){
 	}else {
 		bReturn = false;
@@ -136,7 +136,7 @@ bool CSolidSBCCliConfigSocket::WaitForPacket()
 	//readable socket means packet recieved, 
 	fd_set readfds; FD_ZERO(&readfds); FD_SET(m_hCliConfSocket,&readfds);
 	
-	int nErr = select(m_hCliConfSocket+1,&readfds,NULL,NULL,NULL);
+	int nErr = select((int)(m_hCliConfSocket+1),&readfds,NULL,NULL,NULL);
 	if ( ( nErr == 1) && FD_ISSET(m_hCliConfSocket,&readfds) ){
 		return true;}
 	else {
